Introduce enum Fold

+++ src/main/java/power/Base.java
@@ -7,16 +7,16 @@ import java.util.Map;
 public class Base {
 
     private final BigInteger base;
-    private final Map<Integer, BigInteger> powers = new HashMap<>();
+    private final Map<Long, BigInteger> powers = new HashMap<>();
 
     Base(BigInteger base) {
         this.base = base;
-        powers.put(0, BigInteger.ONE);
-        powers.put(1, base);
-        powers.put(2, base.pow(2));
+        powers.put(0L, BigInteger.ONE);
+        powers.put(1L, base);
+        powers.put(2L, base.pow(2));
     }
 
-    public BigInteger pow(int exp) {
+    public BigInteger pow(long exp) {
         if (exp < 0) throw new IllegalArgumentException("exp cannot be negative");
 
         return powers.computeIfAbsent(exp, n -> isEven(n) ?
@@ -24,7 +24,7 @@ public class Base {
                 base.multiply(Power.base(pow(2)).pow(n / 2)));
     }
 
-    private boolean isEven(int n) {
+    private boolean isEven(long n) {
         return n % 2 == 0;
     }
 }
